+++ src/jllvm/vm/JNIImplementation.hpp
@@ -1,0 +1,12 @@
+
+#pragma once
+
+#include <jni.h>
+
+#include "VirtualMachine.hpp"
+
+namespace jllvm
+{
+/// Returns the 'VirtualMachine' instance associated with the 'JNIEnv'.
+VirtualMachine& virtualMachineFromJNIEnv(JNIEnv* env);
+} // namespace jllvm
