+++ src/jllvm/vm/NativeImplementation.cpp
@@ -1,39 +1,13 @@
 #include "NativeImplementation.hpp"
 
-#include <jllvm/unwind/Unwinder.hpp>
-
-jllvm::VirtualMachine& jllvm::detail::virtualMachineFromJNIEnv(JNIEnv* env)
-{
-    return *reinterpret_cast<jllvm::VirtualMachine*>(env->functions->reserved0);
-}
+#include <jllvm/vm/native/JDK.hpp>
+#include <jllvm/vm/native/Lang.hpp>
 
 void jllvm::registerJavaClasses(VirtualMachine& virtualMachine)
 {
+    using namespace lang;
+    using namespace jdk;
+
     addModels<ObjectModel, ClassModel, ThrowableModel, FloatModel, DoubleModel, SystemModel, ReflectionModel, CDSModel,
               UnsafeModel>(virtualMachine);
-}
-
-const jllvm::ClassObject* jllvm::ReflectionModel::getCallerClass(VirtualMachine& virtualMachine,
-                                                                 GCRootRef<ClassObject> classObject)
-{
-    const ClassObject* result = nullptr;
-    unwindStack(
-        [&](UnwindFrame frame)
-        {
-            std::uintptr_t fp = frame.getFunctionPointer();
-            std::optional<JavaMethodMetadata> data = virtualMachine.getJIT().getJavaMethodMetadata(fp);
-            if (!data)
-            {
-                return UnwindAction::ContinueUnwinding;
-            }
-
-            if (data->classObject == classObject)
-            {
-                return UnwindAction::ContinueUnwinding;
-            }
-            // TODO: If the method has the Java annotation '@CallerSensitive' it should be skipped by this method.
-            result = data->classObject;
-            return UnwindAction::StopUnwinding;
-        });
-    return result;
 }
